+++ src/node/server/node.mjs
@@ -7,10 +7,11 @@ const certSuffix = process.env.CI ? '-ci' : '-dev'
 const certPath = path.join(root, `.config/certificates/ssl${certSuffix}.crt`)
 const keyPath = path.join(root, `.config/certificates/ssl${certSuffix}.key`)
 
-console.log('Using SSL certificate:', certPath)
-
-process.env.NITRO_SSL_CERT = fs.readFileSync(certPath, 'utf8')
-process.env.NITRO_SSL_KEY = fs.readFileSync(keyPath, 'utf8')
+if (fs.existsSync(certPath) && fs.existsSync(keyPath)) {
+  console.log('Using SSL certificate:', certPath)
+  process.env.NITRO_SSL_CERT = fs.readFileSync(certPath, 'utf8')
+  process.env.NITRO_SSL_KEY = fs.readFileSync(keyPath, 'utf8')
+}
 
 // await import(path.join(root, '.output/server/sentry.server.config.mjs'))
 await import(path.join(root, 'playground/.output/server/index.mjs'))