+++ tests/e2e/specs/pages/root/error-404.spec.ts
@@ -5,7 +5,12 @@ import { testPageLoad, testVisualRegression } from '#tests/e2e/utils/tests'
 
 const PAGE_PATH = '/does-not-exist'
 
-// testOgImage(PAGE_PATH)
+// testOgImage(
+// {
+//   en: `a_Social+Preview+Image+for+a+Vio+webpage.,c_Nuxt.takumi,description_Vio+is+%40dargmuesli's+Nuxt+layer.,title_404+-+Not+Found,q_e30,p_Ii9kb2VzLW5vdC1leGl0Ig.png`,
+//   de: `a_Social+Preview+Image+f%C3%BCr+eine+Vio+Webseite.,c_Nuxt.takumi,description_Vio+ist+%40dargmueslis+Nuxt+layer.,title_404+-+Nicht+gefunden,q_e30,p_Ii9kZS9kb2VzLW5vdC1leGl0Ig.png`,
+// }
+// )
 testPageLoad(PAGE_PATH, 404)
 testVisualRegression(PAGE_PATH, true)
 
@@ -28,10 +33,14 @@ vioTest.describe('visual regression', () => {
   vioTest('generates the open graph image', async ({ page }) => {
     if (process.env.VIO_SERVER === 'static') return
 
-    await page.goto(`/__og-image__/image/does-not-exist/og.png`)
-    await expect(page).toHaveScreenshot({ fullPage: true })
+    await page.goto(
+      `/_og/d/a_Social+Preview+Image+for+a+Vio+webpage.,c_Nuxt.takumi,description_Vio+is+%40dargmuesli's+Nuxt+layer.,title_404+-+Not+Found,q_e30,p_Ii9kb2VzLW5vdC1leGl0Ig.png`,
+    )
+    await expect(page).toHaveScreenshot()
 
-    await page.goto(`/__og-image__/image/de/does-not-exist/og.png`)
-    await expect(page).toHaveScreenshot({ fullPage: true })
+    await page.goto(
+      `/_og/d/a_Social+Preview+Image+f%C3%BCr+eine+Vio+Webseite.,c_Nuxt.takumi,description_Vio+ist+%40dargmueslis+Nuxt+layer.,title_404+-+Nicht+gefunden,q_e30,p_Ii9kZS9kb2VzLW5vdC1leGl0Ig.png`,
+    )
+    await expect(page).toHaveScreenshot()
   })
 })