+++ tests/e2e/specs/pages/root/privacy-policy.spec.ts
@@ -7,5 +7,14 @@ import {
 const PAGE_PATH = '/privacy-policy'
 
 testPageLoad(PAGE_PATH)
-testOgImage(PAGE_PATH)
+testOgImage({
+  dynamic: {
+    de: 'a_Social+Preview+Image+für+eine+Vio+Webseite.,c_Nuxt.satori,description_Vio+ist+@dargmueslis+Nuxt+layer.,title_Datenschutzerklärung,p_Ii9kZS9wcml2YWN5LXBvbGljeSI.png',
+    en: `a_Social+Preview+Image+for+a+Vio+webpage.,c_Nuxt.satori,description_Vio+is+@dargmuesli's+Nuxt+layer.,title_Privacy+Policy,p_Ii9wcml2YWN5LXBvbGljeSI.png`,
+  },
+  static: {
+    de: 'o_nnlr68.png',
+    en: `o_unn6zg.png`,
+  },
+})
 testVisualRegression(PAGE_PATH)