+++ tests/e2e/utils/constants.ts
@@ -20,7 +20,7 @@ export const PAGE_READY = async ({
   }
 
   // if (
-  //   process.env.NODE_ENV === 'development' &&
+  //   process.env.VIO_SERVER === 'dev' &&
   //   (!options || options.devTools !== false)
   // ) {
   //   await expect(
@@ -36,5 +36,5 @@ export const PAGE_READY = async ({
   }
 }
 export const SITE_URL =
-  process.env.NUXT_PUBLIC_SITE_URL ||
-  `https://${process.env.HOST || 'localhost'}:${process.env.PORT || '3000'}`
+  process.env.NUXT_PUBLIC_I18N_BASE_URL ||
+  `https://${process.env.HOST || 'app.localhost'}:${process.env.PORT || '3000'}`